[Tech] changes after code review
