+++ backend/src/main/kotlin/fr/gouv/cacem/monitorenv/domain/entities/VesselTypeEnum.kt
@@ -1,0 +1,8 @@
+package fr.gouv.cacem.monitorenv.domain.entities
+
+enum class VesselTypeEnum {
+    FISHING,
+    SAILING,
+    MOTOR,
+    COMMERCIAL,
+}
