--- backend/src/main/kotlin/fr/gouv/cacem/monitorenv/domain/entities/mission/envAction/envActionControl/infraction/VesselTypeEnum.kt
@@ -1,8 +1,0 @@
-package fr.gouv.cacem.monitorenv.domain.entities.mission.envAction.envActionControl.infraction
-
-enum class VesselTypeEnum {
-    FISHING,
-    SAILING,
-    MOTOR,
-    COMMERCIAL,
-}
