+++ backend/src/main/kotlin/fr/gouv/cacem/monitorenv/domain/entities/reporting/TargetDetailsEntity.kt
@@ -1,5 +1,7 @@
 package fr.gouv.cacem.monitorenv.domain.entities.reporting
 
+import fr.gouv.cacem.monitorenv.domain.entities.VesselTypeEnum
+
 data class TargetDetailsEntity(
     val mmsi: String? = null,
     val imo: String? = null,
@@ -7,5 +9,5 @@ data class TargetDetailsEntity(
     val vesselName: String? = null,
     val operatorName: String? = null,
     val size: Int? = null,
-    val vesselType: String? = null,
+    val vesselType: VesselTypeEnum? = null,
 )
